Add new CacheManger class to handle cache checking & updating
