+++ Backend/package.json
@@ -26,6 +26,7 @@
     "body-parser": "^1.19.0",
     "cheerio": "^1.0.0-rc.3",
     "cors": "^2.8.5",
+    "cron-parser": "^2.11.0",
     "express": "^4.16.4",
     "google-libphonenumber": "^3.2.2",
     "helmet": "^3.16.0",
