+++ Backend/src/server.ts
@@ -1,10 +1,16 @@
-// created based on: https://blog.risingstack.com/building-a-node-js-app-with-typescript-tutorial/
-// important: https://docs.npmjs.com/misc/scripts [https://gyazo.com/6e843ae22f1119850a205bca7efbf90a]
 import app from "./App";
+import MAIN_CONFIG from "./config/Main";
+import CacheManger from "./classes/CacheManager";
 
 (app.express).listen(app.port, (err: any) => {
   if(err) throw err;
-  console.log(`Server is listening on ${app.port}`)
+  console.log(`Server is listening on ${app.port}`);
 
-  return app.initCacheChecking();
+  if(process.env && process.env.pm_id) {
+    if(process.env.NODE_APP_INSTANCE === "0") {
+      new CacheManger(MAIN_CONFIG.cacheCheckInterval); // start cache checking and updating (if main PM2 instance)
+    }
+  } else {
+    new CacheManger(MAIN_CONFIG.cacheCheckInterval); // start cache checking and updating
+  }
 });
